Refactor CLI commands
