+++ backend/Makefile
@@ -26,7 +26,7 @@ format:
 	cargo fmt
 
 serve:
-	cargo run
+	cargo run -- serve
 
 sqlite:
 	sqlite3 $(DATABASE)
