+++ backend/src/cli/mod.rs
@@ -1,3 +1,9 @@
-mod upload_files;
+mod osm_pull_command;
+mod queue_consumer_command;
+mod serve_command;
+mod upload_files_command;
 
-pub use self::upload_files::*;
+pub use self::osm_pull_command::*;
+pub use self::queue_consumer_command::*;
+pub use self::serve_command::*;
+pub use self::upload_files_command::*;
