+++ container/rootfs/etc/supervisor.d/queue.ini
@@ -1,5 +1,5 @@
 [program:queue]
-command = bin/treemap --queue-consumer
+command = bin/treemap queue-consumer
 directory = /app
 autostart = true
 autorestart = true
