+++ container/rootfs/etc/supervisor.d/web.ini
@@ -1,5 +1,5 @@
 [program:web]
-command = bin/treemap
+command = bin/treemap serve
 directory = /app
 autostart = true
 autorestart = true
